move affiliates constants directory
