--- src/constants/affiliates.ts
@@ -1,24 +1,0 @@
-export const DEFAULT_BPF = '0';
-
-export const AFFILIATES = {
-  cosmos: [
-    {
-      chainId: 'osmosis-1',
-      affiliate: [
-        {
-          address: '',
-          basis_points_fee: DEFAULT_BPF,
-        },
-      ],
-    },
-    {
-      chainId: 'neutron-1',
-      affiliate: [
-        {
-          address: '',
-          basis_points_fee: DEFAULT_BPF,
-        },
-      ],
-    },
-  ],
-};
