+++ src/constants/skip.ts
@@ -1,1 +1,26 @@
 export const SKIP_BASE_URL = 'https://api.skip.money';
+
+export const DEFAULT_BPF = '0';
+
+export const AFFILIATES = {
+  cosmos: [
+    {
+      chainId: 'osmosis-1',
+      affiliate: [
+        {
+          address: '',
+          basis_points_fee: DEFAULT_BPF,
+        },
+      ],
+    },
+    {
+      chainId: 'neutron-1',
+      affiliate: [
+        {
+          address: '',
+          basis_points_fee: DEFAULT_BPF,
+        },
+      ],
+    },
+  ],
+};
